requirements.txt
# Make sure to install those as additional_dependencies in the
# pre-commit hooks for pylint & mypy
# Also update the readme accordingly

# When dependencies release new versions and tests succeed, we should try to expand the allowed
# versions and only increase the lower bound if necessary

# httpx has no stable release yet, so let's be cautious for now
# HTTP/2 is more performant and stable than HTTP/1.1, specially for concurrent requests
httpx[http2] ~= 0.23.3
python-telegram-bot[ext] == 20.1

ImageHash~=4.3.1
Pillow~=9.4.0
hexhamming~=2.2.3